precision

+++ src/ft_fnormalize.c
@@ -6,7 +6,7 @@
 /*   By: lroussel <lroussel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/16 17:09:17 by lroussel          #+#    #+#             */
-/*   Updated: 2025/05/17 14:28:00 by lroussel         ###   ########.fr       */
+/*   Updated: 2025/05/18 00:10:53 by lroussel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ t_fvector3	ft_fnormalize(t_fvector3 v)
 {
 	float	len;
 
-	len = sqrt(v.x * v.x + v.y * v.y + v.z * v.z);
+	len = sqrtf(v.x * v.x + v.y * v.y + v.z * v.z);
 	if (len == 0.0f)
 	{
 		v.x = 0.0f;
